search: use non-tracking session by default
Before a user would have to always pass a session:

 session := sajari.NewSession(sajari.TrackingNone, "", nil)
 pipeline.Search(context.Background(), values, session)

With this change the user can now omit the session option and Sajari
will default to using it:

 pipeline.Search(context.Background(), values)

If the user wants to provide their own session they can use the
WithSession option:

 session := sajari.NewSession(sajari.TrackingClick, "url", map[string]string{})
 pipeline.Search(context.Background(), values, sajari.WithSession(session))

Perhaps a natural benefit of this change is that it allows the Search
method to allow options in the future in a backwards compatible way.

I considered naming the option PipelineOpt rather than SearchOpt and
that is still possible, but SearchOpt seemed to make more sense.